# openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PoE2 Flipper API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create Event",
        "description": "Create a new manual event flag.\n\nFrom spec section 6, events affect multiple subsystems:\n- SARIMA forecasts: labeled low_confidence=True\n- Holt-Winters: disabled entirely\n- Flip scoring: affected currencies get penalty or exclusion\n- Phase reset: major_patch events reset the PhaseDetector\n\nEvents auto-expire after a configurable duration (default: 48 hours).",
        "operationId": "create_event_api_v1_events_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Events",
        "description": "List all events, optionally filtering to active ones only.",
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return active events",
              "default": true,
              "title": "Active Only"
            },
            "description": "Only return active events"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/anomalies": {
      "get": {
        "tags": [
          "anomalies"
        ],
        "summary": "Get Anomalies",
        "description": "Detect anomalies across all currencies or a specific one.\n\nUses cached historical data from Poe2ScoutProvider and the existing\nAnomalyDetector ensemble (5 indicators: Z-score, MACD, RSI, STL, momentum).\n\nPERFORMANCE: The CPU-bound computation is offloaded to ProcessPoolExecutor\nvia loop.run_in_executor() with asyncio.wait_for(timeout=45s) to prevent\nindefinite blocking. This ensures health check endpoints remain responsive\nduring heavy computation, preventing circuit breaker cascade failures.\n\nArgs:\n    currency: Optional currency API ID to check. If None, checks all monitored currencies.\n    min_alert_score: Minimum alert score to include in results (default 0.4).",
        "operationId": "get_anomalies_api_v1_anomalies_get",
        "parameters": [
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specific currency API ID, or None for all",
              "title": "Currency"
            },
            "description": "Specific currency API ID, or None for all"
          },
          {
            "name": "min_alert_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Minimum alert score to include",
              "default": 0.4,
              "title": "Min Alert Score"
            },
            "description": "Minimum alert score to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomaliesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/batch": {
      "post": {
        "tags": [
          "batch"
        ],
        "summary": "Batch Endpoint",
        "description": "Execute multiple API requests in a single HTTP call.\n\nThis endpoint is designed for initial page load optimization.\nInstead of making 5-6 separate HTTP requests to fetch health,\nphase, events, currencies, and optimal-currency data, the\nfrontend sends a single POST /api/batch request.\n\nAll sub-requests are executed concurrently via asyncio.gather.\nFailed sub-requests are reported in the 'errors' dict without\naffecting successful ones.\n\nThe sub-requests are dispatched as internal HTTP calls to\nlocalhost:8000 (the same FastAPI app). This approach:\n- Reuses all existing route logic without code duplication\n- Respects all existing middleware, error handling, and caching\n- Is safe because it's internal-only (no external network access)",
        "operationId": "batch_endpoint_api_v1_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnomaliesResponse": {
        "properties": {
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyAlertData"
            },
            "type": "array",
            "title": "Anomalies"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of anomalies found"
          },
          "currencies_checked": {
            "type": "integer",
            "title": "Currencies Checked",
            "description": "Number of currencies checked"
          },
          "min_alert_score": {
            "type": "number",
            "title": "Min Alert Score",
            "description": "Minimum alert score filter applied"
          },
          "data_available": {
            "type": "boolean",
            "title": "Data Available",
            "description": "Whether data is available"
          }
        },
        "type": "object",
        "required": [
          "count",
          "currencies_checked",
          "min_alert_score",
          "data_available"
        ],
        "title": "AnomaliesResponse",
        "description": "Response for GET /api/v1/anomalies."
      },
      "AnomalyAlertData": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Currency API identifier"
          },
          "alert_score": {
            "type": "number",
            "title": "Alert Score",
            "description": "Anomaly alert score (0-1)"
          },
          "triggered_indicators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Triggered Indicators",
            "description": "List of triggered indicator names"
          },
          "direction": {
            "type": "string",
            "title": "Direction",
            "description": "Anomaly direction: spike_up, spike_down"
          },
          "is_confirmed": {
            "type": "boolean",
            "title": "Is Confirmed",
            "description": "Whether the anomaly is confirmed by multiple indicators"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Anomaly detection timestamp (ISO 8601)"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "alert_score",
          "triggered_indicators",
          "direction",
          "is_confirmed",
          "timestamp"
        ],
        "title": "AnomalyAlertData",
        "description": "Single anomaly alert."
      },
...
      "BatchRequest": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/BatchSubRequest"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Requests",
            "description": "List of sub-requests to execute"
          }
        },
        "type": "object",
        "required": [
          "requests"
        ],
        "title": "BatchRequest",
        "description": "Top-level batch request body."
      },
      "BatchResponse": {
        "properties": {
          "results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Results"
          },
          "errors": {
            "additionalProperties": true,
            "type": "object",
            "title": "Errors"
          },
          "timing_ms": {
            "type": "number",
            "title": "Timing Ms",
            "description": "Total batch execution time in ms",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "BatchResponse",
        "description": "Batch response containing results and errors."
      },
      "BatchSubRequest": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "title": "Id",
            "description": "Client-defined identifier for this sub-request result lookup"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "API path, e.g. /api/v1/phase"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params",
            "description": "Optional query parameters for the sub-request"
          }
        },
        "type": "object",
        "required": [
          "id",
          "path"
        ],
        "title": "BatchSubRequest",
        "description": "A single sub-request within a batch."
      },
...
      "CreateEventRequest": {
        "properties": {
          "event_type": {
            "$ref": "#/components/schemas/EventType",
            "description": "Type of event: major_patch, minor_patch, league_start, economy_shift, streamer_hype, other"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Description",
            "description": "Human-readable description of the event"
          },
          "affected_currencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Currencies",
            "description": "Optional list of currency API IDs affected by this event. If empty, the event affects all currencies."
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "ISO 8601 timestamp of when the event occurred. Defaults to now."
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "ISO 8601 timestamp of when the event expires. Defaults to now + config.events.default_expiry_hours."
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "description"
        ],
        "title": "CreateEventRequest",
        "description": "Request body for creating a new event."
      },
...
      "EventCreateResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Success message"
          },
          "event": {
            "$ref": "#/components/schemas/EventData",
            "description": "Created event data"
          }
        },
        "type": "object",
        "required": [
          "message",
          "event"
        ],
        "title": "EventCreateResponse",
        "description": "Response for POST /api/v1/events."
      },
      "EventData": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id",
            "description": "Unique event identifier"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "description": "Event type: major_patch, minor_patch, league_start, economy_shift, streamer_hype, other"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Human-readable event description"
          },
          "affected_currencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Currencies"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Event timestamp (ISO 8601)"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "Expiry timestamp (ISO 8601)"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the event is still active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Creation timestamp (ISO 8601)"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "description",
          "timestamp",
          "is_active"
        ],
        "title": "EventData",
        "description": "Single market event."
      },
...
      "EventType": {
        "type": "string",
        "enum": [
          "major_patch",
          "minor_patch",
          "league_start",
          "economy_shift",
          "streamer_hype",
          "other"
        ],
        "title": "EventType",
        "description": "Manual event types that can affect market behaviour."
      },
      "EventsListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of events"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventData"
            },
            "type": "array",
            "title": "Events"
          },
          "fetched_at": {
            "type": "string",
            "title": "Fetched At",
            "description": "ISO 8601 timestamp of data fetch"
          }
        },
        "type": "object",
        "required": [
          "total",
          "fetched_at"
        ],
        "title": "EventsListResponse",
        "description": "Response for GET /api/v1/events."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}